requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==4.0.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
reportlab==4.0.8
whitenoise==6.6.0
django-admin-interface==0.28.1
django-admin-list-filter-dropdown==1.0.3
matplotlib==3.9.1
django-admin-rangefilter==0.13.0
django-admin-charts
fpdf2==2.8.2
django-environ==0.12.0
pandas==2.2.3
openpyxl==3.1.5

